refactor(mdslide-docs): modularize icon files and convert icons/index.tsx to a barrel file
