+++ mdslide-docs/src/components/icons/CopyIcon.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+// copy icon. 
+export const CopyIcon = ({ className }: { className?: string }): React.ReactElement => (
+  <svg className={className} width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+    <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
+    <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
+  </svg>
+);
