+++ mdslide-docs/src/components/icons/ExternalLinkIcon.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+// external link indicator arrow icon. 
+export const ExternalLinkIcon = (): React.ReactElement => (
+  <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" style={{ marginLeft: '4px', verticalAlign: 'middle' }}>
+    <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6" />
+    <polyline points="15 3 21 3 21 9" />
+    <line x1="10" y1="14" x2="21" y2="3" />
+  </svg>
+);
