+++ mdslide-docs/src/components/icons/MathIcon.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+// math symbols hashtag/grid layout feature icon.  
+export const MathIcon = ({ className = 'w-6 h-6 stroke-app-accent stroke-[1.5] fill-none' }: { className?: string }): React.ReactElement => (
+  <svg className={className} viewBox="0 0 24 24" strokeLinecap="round" strokeLinejoin="round">
+    <line x1="4" y1="9" x2="20" y2="9" />
+    <line x1="4" y1="15" x2="20" y2="15" />
+    <line x1="10" y1="3" x2="8" y2="21" />
+    <line x1="16" y1="3" x2="14" y2="21" />
+  </svg>
+);
