+++ mdslide-docs/src/components/icons/MoonIcon.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// moon crescent icon 
+export const MoonIcon = (): React.ReactElement => (
+  <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+    <path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z" />
+  </svg>
+);
